silence linux warning

+++ libretro-common/net/net_http.c
@@ -329,13 +329,13 @@ struct http_t *net_http_new(struct http_connection_t *conn)
 
       post_len = strlen(conn->postdatacopy);
 #ifdef _WIN32
-      len = snprintf(NULL, 0, "%I64u", (long long unsigned)post_len);
+      len = snprintf(NULL, 0, "%I64u", post_len);
       len_str = (char*)malloc(len + 1);
-      snprintf(len_str, len + 1, "%I64u", (long long unsigned)post_len);
+      snprintf(len_str, len + 1, "%I64u", post_len);
 #else
-      len = snprintf(NULL, 0, "%llu", post_len);
+      len = snprintf(NULL, 0, "%llu", (long long unsigned)post_len);
       len_str = (char*)malloc(len + 1);
-      snprintf(len_str, len + 1, "%llu", post_len);
+      snprintf(len_str, len + 1, "%llu", (long long unsigned)post_len);
 #endif
 
       len_str[len] = '\0';
